refactor(suggester): move tokenStringInspector to a specific file
